+++ src/config.ts
@@ -1,0 +1,34 @@
+import { Disposable, commands } from 'vscode';
+import { SELECT_ENV_COMMAND_ID } from './utilities/consts';
+import { selectEnvPreset } from './command_implementations/selectEnvPreset';
+import { FileSystemHandler } from './handlers';
+import { selectedEnvPresetEventEmitter } from './utilities/events';
+import EnvStatusBarItem from './status_bar_items/envStatusBarItem';
+import { EnvHandler } from './handlers/envHandler';
+import { PresetChangeWatcher } from './watchers/presetChangeWatcher';
+
+interface InitializationDeps {
+  subscriptions: Disposable[];
+}
+
+export async function initialize({ subscriptions }: InitializationDeps) {
+  const fsHandler = new FileSystemHandler();
+  const envHandler = await EnvHandler.build({ fsHandler });
+
+  const selectEnvPresetCmd = commands.registerCommand(SELECT_ENV_COMMAND_ID, () =>
+    selectEnvPreset({
+      envHandler,
+      rootDir: fsHandler.rootDir.uri,
+    }),
+  );
+  const statusBarItem = await EnvStatusBarItem.build({ envHandler });
+  const presetChangeWatcher = new PresetChangeWatcher({
+    rootDir: fsHandler.rootDir,
+    envHandler,
+  });
+
+  subscriptions.push(selectedEnvPresetEventEmitter);
+  subscriptions.push(selectEnvPresetCmd);
+  subscriptions.push(statusBarItem);
+  subscriptions.push(presetChangeWatcher);
+}